Refactoring - Rename file : 색약 색상만 반환 - Delete Unused Function : color-convert 사용으로 unused - Color Table rgb 값으로 변경 - const/let 정리
